+++ Shared/EventHandler.cs
@@ -35,11 +35,13 @@ namespace Slipstream.Shared
         public delegate void OnInternalCommandPluginStatesHandler(EventHandler source, EventHandlerArgs<Shared.Events.Internal.InternalCommandPluginStates> e);
         public event OnInternalCommandPluginStatesHandler? OnInternalCommandPluginStates;
 
-        public delegate void OnInternalReconfiguredHandler(EventHandler source, EventHandlerArgs<Shared.Events.Internal.InternalReconfigured> e);
-        public event OnInternalReconfiguredHandler? OnInternalReconfigured;
+        public delegate void OnInternalCommandReconfigureHandler(EventHandler source, EventHandlerArgs<Shared.Events.Internal.InternalCommandReconfigure> e);
+        public event OnInternalCommandReconfigureHandler? OnInternalCommandReconfigure;
+        #endregion
 
-        public delegate void OnInternalBootupEventsHandler(EventHandler source, EventHandlerArgs<Shared.Events.Internal.InternalBootupEvents> e);
-        public event OnInternalBootupEventsHandler? OnInternalBootupEvents;
+        #region Events: LuaManager
+        public delegate void OnInternalCommandDeduplicateEventsHandler(EventHandler source, EventHandlerArgs<Shared.Events.LuaManager.LuaManagerCommandDeduplicateEvents> e);
+        public event OnInternalCommandDeduplicateEventsHandler? OnLuaManagerCommandDeduplicateEvents;
 
         #endregion
 
@@ -201,17 +203,20 @@ namespace Slipstream.Shared
                     else
                         OnInternalCommandPluginStates.Invoke(this, new EventHandlerArgs<Shared.Events.Internal.InternalCommandPluginStates>(tev));
                     break;
-                case Shared.Events.Internal.InternalReconfigured tev:
-                    if (OnInternalReconfigured == null)
+                case Shared.Events.Internal.InternalCommandReconfigure tev:
+                    if (OnInternalCommandReconfigure == null)
                         OnDefault?.Invoke(this, new EventHandlerArgs<IEvent>(tev));
                     else
-                        OnInternalReconfigured.Invoke(this, new EventHandlerArgs<Shared.Events.Internal.InternalReconfigured>(tev));
+                        OnInternalCommandReconfigure.Invoke(this, new EventHandlerArgs<Shared.Events.Internal.InternalCommandReconfigure>(tev));
                     break;
-                case Shared.Events.Internal.InternalBootupEvents tev:
-                    if (OnInternalBootupEvents == null)
+
+                // LuaManager
+
+                case Shared.Events.LuaManager.LuaManagerCommandDeduplicateEvents tev:
+                    if (OnLuaManagerCommandDeduplicateEvents == null)
                         OnDefault?.Invoke(this, new EventHandlerArgs<IEvent>(tev));
                     else
-                        OnInternalBootupEvents.Invoke(this, new EventHandlerArgs<Shared.Events.Internal.InternalBootupEvents>(tev));
+                        OnLuaManagerCommandDeduplicateEvents.Invoke(this, new EventHandlerArgs<Shared.Events.LuaManager.LuaManagerCommandDeduplicateEvents>(tev));
                     break;
 
                 // File Monitor