+++ Shared/IEventFactory.cs
@@ -2,6 +2,7 @@
 using Slipstream.Shared.Events.FileMonitor;
 using Slipstream.Shared.Events.Internal;
 using Slipstream.Shared.Events.IRacing;
+using Slipstream.Shared.Events.LuaManager;
 using Slipstream.Shared.Events.Twitch;
 using Slipstream.Shared.Events.UI;
 
@@ -38,8 +39,8 @@ namespace Slipstream.Shared
         InternalCommandPluginStates CreateInternalCommandPluginStates();
         InternalCommandPluginUnregister CreateInternalCommandPluginUnregister(string pluginId);
         InternalPluginState CreateInternalPluginState(string pluginId, string pluginName, string displayName, PluginStatusEnum pluginStatus);
-        InternalReconfigured CreateInternalReconfigured();
-        InternalBootupEvents CreateInternalBootupEvents(IEvent[] events);
+        InternalCommandReconfigure CreateInternalCommandReconfigure();
+        LuaManagerCommandDeduplicateEvents CreateLuaManagerCommandDeduplicateEvents(IEvent[] events);
 
         IRacingCarCompletedLap CreateIRacingCarCompletedLap(double sessionTime, long carIdx, double time, int lapsCompleted, float? fuelDiff, bool localUser);
         IRacingCarInfo CreateIRacingCarInfo(